requirements.txt
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
bcrypt==4.3.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
passlib==1.7.4
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.42
sqlmodel==0.0.24
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
cryptography